+++ .prettierrc
@@ -1,0 +1,6 @@
+{
+  "bracketSameLine": true,
+  "printWidth": 120,
+  "singleQuote": true,
+  "trailingComma": "none"
+}
